Fix/Correct description function... When the description function was first added to operators, I didn't understand how it worked but finally got a functional result and used it as a template ever since.
Either it was changed at some point or I just never fully got it.  `cls.bl_rna.description` is (or at least the ones I see), is an empty string.

It turns out if the description function returns a False value, it will default to the class' `bl_description`, so  `return cls.bl_rna.description` is equivalent to `return None` and `return ""`.